requirements.txt


instagrapi==2.1.2





instaloader==4.13

requests==2.32.3

discord-py==2.4.0

instagramy==4.5